test(core-unittest): split birthday tests
